Check blank ID in single resource fetch requests (#53)
Whenever we fetch a single request, we need to check that the provided
ID is valid and not let the request go through if the ID is blank.
Since we generate the request endpoints (path) dynamically, we have to
guarantee ID existence.
As an example, let's say we want to fetch a single user and provide a
blank id (''). The generated path will be '/users/' + ''. If this
request goes through, it will fetch the list of users instead.

Added a requireId() protected function in AbstractApi and included the
check in all single resource fetching methods in the SDK.