fix: Properly read header in cross-origin check
Since header field names are case-insensitive, Node transforms all
headers names to lowercase. For example, if a request sets the header
as:

    X-Forwarded-Host: example.com

Node will transform it to:

    x-forwarded-host: example.com

Therefore, the cross-origin detection algorithm was broken with regards
to reverse proxies, because it was reading the wrong header,
X-Forwarded-Host, which was always undefined.

First reported in https://discord.com/channels/856971667393609759/858012732677947402/924739777557827695